requirements.txt
cryptography==37.0.1
Flask==2.1.0
PyMySQL==1.0.2
python-dotenv==1.0.0
requests==2.28.1
